+++ server/v2/core/event/event.go
@@ -1,9 +1,5 @@
 package event
 
-import (
-	"cosmossdk.io/core/event"
-)
-
 // Events represents a list of events.
 type Events struct {
 	Events []Event
@@ -12,5 +8,10 @@ type Events struct {
 // Event defines how an event will emitted
 type Event struct {
 	Type       string
-	Attributes []event.Attribute
+	Attributes []Attribute
+}
+
+// KVEventAttribute is a kv-pair event attribute.
+type Attribute struct {
+	Key, Value string
 }
