+++ server/v2/core/store/types.go
