+++ server/v2/grpc/server.go
@@ -10,10 +10,10 @@ import (
 
 	"cosmossdk.io/log"
 
+	"cosmossdk.io/server/v2/grpc/gogoreflection"
+	reflection "cosmossdk.io/server/v2/grpc/reflection/v2alpha1"
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/codec"
-	"github.com/cosmos/cosmos-sdk/serverv2/grpc/gogoreflection"
-	reflection "github.com/cosmos/cosmos-sdk/serverv2/grpc/reflection/v2alpha1"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	_ "github.com/cosmos/cosmos-sdk/types/tx/amino" // Import amino.proto file for reflection
 )
