+++ server/v2/stf/go.mod
@@ -1,6 +1,6 @@
 module cosmossdk.io/server/v2/stf
 
-go 1.21.1
+go 1.21
 
 replace cosmossdk.io/server/v2/core => ../core
