+++ server/v2/stf/types.go
