+++ server/v2/transaction/valdiation.go
@@ -3,7 +3,7 @@ package transaction
 import (
 	"context"
 
-	tx "github.com/cosmos/cosmos-sdk/serverv2/core/transaction"
+	tx "cosmossdk.io/server/v2/core/transaction"
 )
 
 var _ tx.Validator[tx.Tx] = TxValidator[tx.Tx]{}
