--- serverv2/transaction/errors.go
@@ -1,5 +1,0 @@
-package transaction
-
-import "errors"
-
-var ErrorInvalidTransaction = errors.New("invalid transaction")
